New Dynamic Model pattern to support type-checked additional properties
Why:
The SDK code generator is being updated to improve the pattern used for
dynamic models (i.e. models that support "additionalProperties").

What:
The changes in this PR can be summarized as follows:
1) An enhanced DynamicModel class which is now parameterized with the
additionalProperties value type and includes a map for them
2) A new DynamicModelTypeAdapterFactory class which is registered with Gson
to perform serialization and deserialization of ALL generated DynamicModel
subclasses.
3) GsonSerializationHelper is now deprecated as it's no longer needed.
4) Tests related to the above functional changes.