feat: RequestBuilder changes to support JSON streaming feature
Fixes arf/planning-sdk-squad#901

The RequestBuilder.bodyContent() method is modified to no
longer use the contentType parameter to determine which one of the
other three input parameters should be used to create the request body.
Instead, we now use the first non-null value using the order
'jsonContent', 'jsonPatchContent', then 'nonJsonContent'.

This change is necessary for us to implement the JSON streaming
feature, where the SDK user might serialize the JSON content
himself and pass it in as a String or InputStream while also
using a contentType that would otherwise indicate a JSON body.