fix: support serialization of models w/lists of discriminated subclasses
Fixes arf/planning-sdk-squad#2011

We initially did not implement the "write()" method within the
DiscriminatorBasedTypeAdapterFactory because we should never
encounter an actual instance of a discriminated base class
(they are implemented as abstract base classes).

However, in some scenarios Gson will use a type adapter
that reflects the DECLARED type rather than the actual RUNTIME type
associated with an object.  One such scenario is when a class
contains a field which is defined as a List of some type X, where
X is a discriminated base class (oneOf parent).

For this reason, it's possible that the write() method is
called.  This commit implements the write() method so that it
simply delegates to the TypeAdapter associated with the
object's RUNTIME type.