fix: improve date and date-time serialization/deserialization
Fixes arf/planning-sdk-squad#2291

This commit contains a re-alignment of our support for
serialization and de-serialization of java.util.Date values
within the Java core:
1. The basic function of formatting and parsing java.util.Date
values as an OpenAPI "date" or "date-time" value is now isolated
in the DateUtils class.
2. The Gson serialization/deserialization of "date-time" values is now
implemented in the DateTimeTypeAdapter class. This class is also the
default type adapter registered with Gson for java.util.Date instances.
3. The Gson serialization/deserialization of "date" values is
implemented in the DateTypeAdapter class.  This class can be used
with the @JsonAdapter annotation to override the ser/deser handling
of a java.util.Date field that should be interpretted as a "date" value.
This will be done by the Java generator.