feat: wrap exceptions during client-side response processing
This commit introduces a new InvalidServiceResponseException class,
which will wrap any exception caught while processing the response
received from the server for an otherwise successful operation
invocation.  For example, an exception thrown by GSON due to a
JSON unmarshalling error will now be wrapped in an
InvalidServiceResponseException (a subclass of ServiceResponseException).