feat(authentication): support cp4d /v1/authorize operation
The main focus of this change is to enhance the CloudPakForDataAuthentictor
implementation to support the official "POST /v1/authorize" operation for
obtaining an access token.   This includes support for using either a password
or an apikey when obtaining the access token from the token service.

In addition, a new Builder class was introduced which provides a more user-friendly
way to construct instances of CloudPakForDataAuthenticator.  The "fromConfiguration"
method was also added to replace the (now deprecated) map-based ctor.

In order to keep the IamAuthenticator class in sync with CloudPakForDataAuthenticator,
the Builder class and fromConfiguration() method were also added to IamAuthenticator.
In the case of both authenticators, the existing set of constructors are still supported
but are now deprecated and will be removed in the next major release of the Java core.