feat(authenticator): add new cp4d service authenticator
This commit introduces a new authenticator implementation
that supports the "GET /v1/service_token" operation which is part
of the Cloud Pak For Data authentication service.
This allows services deployed in CP4D environments to interact
with other services while using the "server broker secret"
configured in the CP4D environment.