fix(build): publish artifacts directly to maven central repository
This commit modifies the build configuration so that build artifacts
are published to the Maven Central repository (via the ossrh
staging repo).  Previously, build artifacts were published to
the bintray/jcenter repository, then mirrored to Maven Central.
But, with the imminent demise of bintray/jcenter, we need to
publish directly to Maven Central.  Going forward, published jars
will be available only on Maven Central.