feat: support user-defined retry interceptors (#169)
This commit provides a pluggable way for users of the java core
to supply their own retry interceptor implementation in order
to customize the conditions under which retries are automatically
attempted.
Included in these changes is the notion of a "retry strategy",
which is a factory for creating retry interceptor instances.
Users of the java core (e.g. an SDK project) can implement
their own retry strategy and then set it with the
HttpClientSingleton.setRetryStrategy() static method.
Once set, the registered factory's createRetryInterceptor()
method will be called whenever the java core needs to initialize an
http client instance when retries are enabled.
The factory's createRetryInterceptor() method returns an implementation
of the IRetryInterceptor interface (which is itself a subclass of
the okhttp3 Interceptor interface).
User's might find it convenient to define their retry interceptor
class as a subclass of the java core's RetryInterceptor class,
thereby relying on RetryInterceptor for most of the functionality,
while overriding specific methods as needed.